haystack-search-server/requirements.txt
haystack-ai>=2.12.0
qdrant-haystack>=6.0.0
sentence-transformers>=3.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.0.0
